+++ src/app/api/kakao/logout/callback/route.ts
@@ -1,5 +1,5 @@
 import { NextResponse } from 'next/server';
-import { signOut } from '@/app/api/auth/[...nextauth]/route';
+import { signOut } from '@/domain/user/utils/auth';
 
 export async function GET(request: Request) {
   await signOut({ redirect: false });
