+++ src/components/header/LogoutButton.tsx
@@ -6,7 +6,7 @@ import type { SessionType } from 'next-auth';
 import { signOut as nextAuthSignOut } from 'next-auth/react';
 import { useTransition } from 'react';
 import { toast } from '@/components/toast';
-import { getKaKaoLogoutURL } from '@/domain/user/utils/auth';
+import { getKaKaoLogoutURL } from '@/domain/user/utils/kakaoAuth';
 
 export default function LogoutForm({
   sessionType,
