+++ src/middleware.ts
@@ -1,5 +1,5 @@
 import { NextResponse } from 'next/server';
-import { auth } from '@/app/api/auth/[...nextauth]/route';
+import { auth } from '@/domain/user/utils/auth';
 
 export default auth((req) => {
   const baseUrl = req.nextUrl.origin;
