+++ src/utils/getAuth.ts
@@ -1,5 +1,5 @@
 import { getSession } from 'next-auth/react';
-import { auth } from '@/app/api/auth/[...nextauth]/route';
+import { auth } from '@/domain/user/utils/auth';
 
 export async function getAuth() {
   if (typeof window === 'undefined') {
